:sparkles: Add rewrite for bv 1
