+++ src/Grisette/Internal/SymPrim/Prim/Internal/Instances/PEvalRotateTerm.hs
@@ -27,8 +27,6 @@ import qualified Data.SBV as SBV
 import GHC.TypeLits (KnownNat, type (<=))
 import Grisette.Internal.Core.Data.Class.SymRotate (SymRotate (symRotate))
 import Grisette.Internal.SymPrim.BV (IntN, WordN)
-import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalBitwiseTerm ()
-import Grisette.Internal.SymPrim.Prim.Internal.Instances.SupportedPrim (bvIsNonZeroFromGEq1)
 import Grisette.Internal.SymPrim.Prim.Internal.Term
   ( PEvalRotateTerm
       ( pevalRotateLeftTerm,
@@ -39,6 +37,7 @@ import Grisette.Internal.SymPrim.Prim.Internal.Term
       ),
     SupportedNonFuncPrim (withNonFuncPrim),
     Term,
+    bvIsNonZeroFromGEq1,
     conTerm,
     rotateLeftTerm,
     rotateRightTerm,
