+++ src/Grisette/Internal/SymPrim/Prim/Internal/Instances/PEvalShiftTerm.hs
@@ -26,10 +26,6 @@ import Data.Proxy (Proxy (Proxy))
 import GHC.TypeLits (KnownNat, type (<=))
 import Grisette.Internal.Core.Data.Class.SymShift (SymShift (symShift))
 import Grisette.Internal.SymPrim.BV (IntN, WordN)
-import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalBitwiseTerm ()
-import Grisette.Internal.SymPrim.Prim.Internal.Instances.SupportedPrim
-  ( bvIsNonZeroFromGEq1,
-  )
 import Grisette.Internal.SymPrim.Prim.Internal.Term
   ( PEvalShiftTerm
       ( pevalShiftLeftTerm,
@@ -39,6 +35,7 @@ import Grisette.Internal.SymPrim.Prim.Internal.Term
     SupportedNonFuncPrim (withNonFuncPrim),
     SupportedPrim,
     Term,
+    bvIsNonZeroFromGEq1,
     conTerm,
     shiftLeftTerm,
     shiftRightTerm,
